+++ Harmonics/CHANGELOG.md
@@ -1,5 +1,23 @@
-﻿# Harmonics Changelog
+﻿# HARMONICS Changelog
+
+## 1.0.2
+
+### Added
+
+* Added `recipe` console command for adding Gemstone ring recipes to the crafting menu, for players already past the required levels. The entry command is `hrmn`. E.g., `hrmn recipe garnet`.
+
+### Fixed
+
+* Added the missing Garnet Ring recipe.
+
+## 1.0.1
+
+### Fixed
+
+* Fixed the absurdly high frequency of Garnet node spawns which had been increased for testing.
+
+<sup><sup>[🔼 Back to top](#harmonics-changelog)</sup></sup>
 
 ## 1.0.0 - Initial 1.6 release
 
-[🔼 Back to top](#Harmonics-changelog)
+[🔼 Back to top](#harmonics-changelog)